+++ example/lib/screens/component/multi_component/multi_component_advanced_screen.dart
@@ -154,8 +154,8 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
       onAdditionalDetails: googlePayRepository.onAdditionalDetails,
     );
 
-    final GooglePayComponentConfiguration googlePayComponentConfiguration =
-        GooglePayComponentConfiguration(
+    final CheckoutConfiguration googlePayCheckoutConfiguration =
+        CheckoutConfiguration(
       environment: Config.environment,
       clientKey: Config.clientKey,
       countryCode: Config.countryCode,
@@ -179,7 +179,7 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.fromLTRB(16, 0, 16, 0),
       child: AdyenGooglePayComponent(
-        configuration: googlePayComponentConfiguration,
+        configuration: googlePayCheckoutConfiguration,
         paymentMethod: paymentMethod,
         checkout: advancedCheckout,
         style: googlePayButtonStyle,
@@ -197,8 +197,8 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
     Map<String, dynamic> paymentMethods,
     BuildContext context,
   ) {
-    final ApplePayComponentConfiguration applePayComponentConfiguration =
-        ApplePayComponentConfiguration(
+    final CheckoutConfiguration applePayCheckoutConfiguration =
+        CheckoutConfiguration(
       environment: Config.environment,
       clientKey: Config.clientKey,
       countryCode: Config.countryCode,
@@ -219,7 +219,7 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.fromLTRB(16, 0, 16, 0),
       child: AdyenApplePayComponent(
-        configuration: applePayComponentConfiguration,
+        configuration: applePayCheckoutConfiguration,
         paymentMethod: paymentMethod,
         checkout: advancedCheckout,
         loadingIndicator: const CircularProgressIndicator(),