+++ lib/src/components/google_pay/google_pay_session_component.dart
@@ -15,7 +15,7 @@ class GooglePaySessionComponent extends BaseGooglePayComponent {
     super.key,
     required this.session,
     required super.googlePayPaymentMethod,
-    required super.googlePayComponentConfiguration,
+    required super.configuration,
     required super.onPaymentResult,
     required super.theme,
     required super.type,